+++ previous-versions/accounting/src/sdk/models/shared/billlineitem.ts
@@ -107,8 +107,6 @@ export class BillLineItem extends SpeakeasyBase {
 
     /**
      * Collection of categories against which this item is tracked.
-     *
-     * @deprecated field: This will be removed in a future release, please migrate away from it as soon as possible.
      */
     @SpeakeasyMetadata({ elemType: TrackingCategoryRef })
     @Expose({ name: "trackingCategoryRefs" })