+++ previous-versions/accounting/src/sdk/models/shared/creditnotelineitem.ts
@@ -3,9 +3,9 @@
  */
 
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
+import { AccountingCustomerRef } from "./accountingcustomerref";
 import { AccountRef } from "./accountref";
 import { BilledToType1 } from "./billedtotype1";
-import { CustomerRef } from "./customerref";
 import { InvoiceTo } from "./invoiceto";
 import { ItemRef } from "./itemref";
 import { ProjectRef } from "./projectref";
@@ -24,8 +24,8 @@ export class CreditNoteLineItemTracking extends SpeakeasyBase {
 
     @SpeakeasyMetadata()
     @Expose({ name: "customerRef" })
-    @Type(() => CustomerRef)
-    customerRef?: CustomerRef;
+    @Type(() => AccountingCustomerRef)
+    customerRef?: AccountingCustomerRef;
 
     @SpeakeasyMetadata()
     @Expose({ name: "isBilledTo" })