+++ previous-versions/accounting/RELEASES.md
@@ -649,3 +649,13 @@ Based on:
 - [typescript v0.34.0] accounting
 ### Releases
 - [NPM v0.34.0] https://www.npmjs.com/package/@codat/accounting/v/0.34.0 - accounting
+
+## 2023-09-01 16:31:46
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.77.1 (2.91.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.35.0] previous-versions/accounting
+### Releases
+- [NPM v0.35.0] https://www.npmjs.com/package/@codat/accounting/v/0.35.0 - previous-versions/accounting